rfact: traits
