+++ src/Domains/Connectors/NetSuite/Services/NetSuiteServices.php
@@ -9,15 +9,12 @@ use Kanvas\Companies\Models\Companies;
 use Kanvas\Connectors\NetSuite\Client;
 use Kanvas\Connectors\NetSuite\DataTransferObject\NetSuite as NetSuiteDto;
 use Kanvas\Connectors\NetSuite\Enums\ConfigurationEnum;
+use Kanvas\Connectors\NetSuite\Traits\UseNetSuiteCustomerSearchTrait;
 use Kanvas\Exceptions\ValidationException;
-use NetSuite\Classes\CustomerSearchBasic;
-use NetSuite\Classes\SearchRequest;
-use NetSuite\Classes\SearchStringField;
-use NetSuite\NetSuiteService as NetSuiteSdkService;
 
 class NetSuiteServices
 {
-    private NetSuiteSdkService $service;
+    use UseNetSuiteCustomerSearchTrait;
 
     public function __construct(
         protected AppInterface $app,
@@ -53,33 +50,5 @@ class NetSuiteServices
         }
 
         return $data->app->set(ConfigurationEnum::NET_SUITE_ACCOUNT_CONFIG->value, $configData);
-    }
-
-    public function findExistingCustomer(string $email): ?object
-    {
-        $searchRequest = new SearchRequest();
-        $searchRequest->searchRecord = $this->createEmailSearchCriteria($email);
-
-        $searchResponse = $this->service->search($searchRequest);
-
-        if ($searchResponse->searchResult->status->isSuccess &&
-            $searchResponse->searchResult->totalRecords > 0) {
-            return $searchResponse->searchResult->recordList->record[0];
-        }
-
-        return null;
-    }
-
-    /**
-     * Create email search criteria for NetSuite.
-     */
-    public function createEmailSearchCriteria(string $email): CustomerSearchBasic
-    {
-        $customerSearch = new CustomerSearchBasic();
-        $customerSearch->email = new SearchStringField();
-        $customerSearch->email->operator = 'is';
-        $customerSearch->email->searchValue = $email;
-
-        return $customerSearch;
     }
 }
